requirements.txt
flask
scikit-learn
pandas
numpy
python-dotenv
requests
Werkzeug
joblib
PyJWT